+++ bluebird/api/__init__.py
@@ -83,11 +83,13 @@ FLASK_API.add_resource(res.Ic, '/ic')
 FLASK_API.add_resource(res.Op, '/op')
 FLASK_API.add_resource(res.Reset, '/reset')
 FLASK_API.add_resource(res.Scenario, '/scenario')
+FLASK_API.add_resource(res.Step, '/step')
 FLASK_API.add_resource(res.Time, '/time')
 
 # Application control
 # FLASK_API.add_resource(res.EpInfo, '/epinfo')
 FLASK_API.add_resource(res.EpLog, '/eplog')
+FLASK_API.add_resource(res.SimMode, '/simmode')
 FLASK_API.add_resource(res.Shutdown, '/shutdown')
 
 # endregion