+++ run.py
@@ -25,6 +25,7 @@ def _parse_args():
 	parser.add_argument('--reset_sim', action='store_true', help='Reset the simulation on '
 	                                                             'connection')
 	parser.add_argument('--log_rate', type=float, help='Log rate in sim-seconds')
+	parser.add_argument('--sim_mode', type=str, help='Set the initial mode')
 	args = parser.parse_args()
 
 	if args.bluesky_host:
@@ -35,6 +36,13 @@ def _parse_args():
 			settings.SIM_LOG_RATE = args.log_rate
 		else:
 			raise ValueError('Rate must be positive')
+
+	mode = args.sim_mode
+	if mode:
+		if not mode in settings.SIM_MODES:
+			available = ', '.join(settings.SIM_MODES)
+			raise ValueError(f'Mode \'{mode}\' not supported. Must be one of: {available}')
+		settings.SIM_MODE = mode
 
 	return args
 
@@ -52,9 +60,10 @@ def main():
 	:return:
 	"""
 
-	args = _parse_args()
 	load_dotenv(verbose=True, override=True)
 	bs_min_version = _get_min_bs_version()
+
+	args = _parse_args()
 
 	with BlueBird() as app:
 		if app.client_connect(bs_min_version, args.reset_sim):