+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump required BlueSky version to `1.3.0-dev`
 - When in `agent` mode, aircraft data is only logged after each `STEP` command
+- Point BlueSky submodule at `tags/turing-1.2.1`
 
 ### Removed